+++ TDEngine2/include/editor/ISelectionManager.h
@@ -26,6 +26,9 @@ namespace TDEngine2
 	class IGraphicsContext;
 
 
+	enum class TTextureHandleId : U32;
+
+
 	TDE2_DECLARE_SCOPED_PTR(IResourceManager)
 	TDE2_DECLARE_SCOPED_PTR(IWindowSystem)
 	TDE2_DECLARE_SCOPED_PTR(IGraphicsContext)
@@ -42,8 +45,6 @@ namespace TDEngine2
 	class ISelectionManager: public virtual IBaseObject, public IEventHandler
 	{
 		public:
-			typedef std::function<void()> TRenderFrameCallback;
-		public:
 			/*!
 				\brief The method initializes the internal state of the manager
 
@@ -58,15 +59,14 @@ namespace TDEngine2
 			TDE2_API virtual E_RESULT_CODE Init(TPtr<IResourceManager> pResourceManager, TPtr<IWindowSystem> pWindowSystem, TPtr<IGraphicsContext> pGraphicsContext, IEditorsManager* pEditorsManager) = 0;
 
 			/*!
-				\brief The method builds so called selection map which is a render target that contains
-				information about which object occupies each pixel of the screen
+				\brief The method is used to update internal selection buffer based on incoming texture with selectables data
 
-				\param[in] onDrawVisibleObjectsCallback A callback in which all visible objects should be drawn
+				\param[in] newSelectionMapHandle A handle that point to ITexture2D instance that hold all the on-screen objects that can be selected
 
 				\return RC_OK if everything went ok, or some other code, which describes an error
 			*/
 
-			TDE2_API virtual E_RESULT_CODE BuildSelectionMap(const TRenderFrameCallback& onDrawVisibleObjectsCallback) = 0;
+			TDE2_API virtual E_RESULT_CODE UpdateSelectionsBuffer(TTextureHandleId newSelectionMapHandle) = 0;
 
 			/*!
 				\brief The method tries to pick an object